backend/requirements.txt
opencv-python
numpy
scikit-image
scikit-learn
elasticsearch
python-dotenv